chore(server): create user leave test file
